Convert to nodelets by default, make direct tag detection an option in launch
