+++ config/perception.yaml
@@ -1,3 +1,5 @@
-grab_resolution: 2
-grab_target_fps: 50
-image_width: 1280
+zed_nodelet:
+  grab_resolution: 2
+  grab_target_fps: 50
+  image_width: 1280
+  image_height: 720
