+++ launch/zed_test.launch
@@ -1,7 +1,8 @@
 <launch>
-    <arg name="run_tag_detector" default="false"/>
     <arg name="run_rviz" default="true"/>
     <arg name="run_dynamic_reconfigure" default="true"/>
+    <arg name="run_tag_detector" default="true"/>
+    <arg name="direct_tag_detection" default="false"/>
 
     <group if="$(arg run_rviz)">
         <arg name="rvizconfig" default="$(find mrover)/config/rviz/zed_test.rviz"/>
@@ -12,14 +13,18 @@
           name="rqt_reconfigure" pkg="rqt_reconfigure" type="rqt_reconfigure"/>
 
     <node pkg="nodelet" type="nodelet" name="perception_nodelet_manager" respawn="true"
-          args="manager" output="screen" />
+          args="manager" output="screen"/>
 
-    <include file="$(find mrover)/launch/zed.launch"/>
+    <include file="$(find mrover)/launch/zed.launch">
+        <arg name="direct_tag_detection" value="$(arg direct_tag_detection)"/>
+    </include>
 
-    <node if="$(arg run_tag_detector)"
-          pkg="nodelet" type="nodelet" name="tag_detector" respawn="true"
-          args="load mrover/TagDetectorNodelet perception_nodelet_manager" output="screen">
-    </node>
+    <group if="$(arg run_tag_detector)">
+        <node unless="$(arg direct_tag_detection)"
+              pkg="nodelet" type="nodelet" name="tag_detector" respawn="true"
+              args="load mrover/TagDetectorNodelet perception_nodelet_manager" output="screen">
+        </node>
+    </group>
 
     <node pkg="tf" type="static_transform_publisher" name="map_odom_publisher"
           args="1 0 0 0 0 0 map odom 100"/>
