+++ src/perception/zed_wrapper/zed_wrapper.cpp
@@ -1,43 +1,32 @@
 #include "zed_wrapper.hpp"
 
-#include <algorithm>
 #include <chrono>
-#include <execution>
 
+#include <nodelet/loader.h>
 #include <ros/init.h>
-#include <sensor_msgs/Imu.h>
-#include <sensor_msgs/image_encodings.h>
-#include <sensor_msgs/point_cloud2_iterator.h>
 
 #include <se3.hpp>
-
-constexpr float DEG2RAD = M_PI / 180.0f;
 
 using namespace std::chrono_literals;
 using hr_clock = std::chrono::high_resolution_clock;
 
 namespace mrover {
 
-    struct Point {
-        float x, y, z;
-        uint8_t r, g, b, a;
-        float normal_x, normal_y, normal_z;
-        float curvature;
-    } __attribute__((packed));
-
-    ZedNode::ZedNode(ros::NodeHandle const& nh, ros::NodeHandle const& pnh)
-        : mNh{nh}, mPnh{pnh}, mTfListener{mTfBuffer}, mIt{mNh},
-          mPcPub{mNh.advertise<sensor_msgs::PointCloud2>("camera/left/points", 1)},
-          mImuPub{mNh.advertise<sensor_msgs::Imu>("imu", 1)},
-          mLeftImgPub{mIt.advertise("camera/left/image", 1)} {
-
+    void ZedNodelet::onInit() {
         try {
+            mNh = getMTNodeHandle();
+            mPnh = getMTPrivateNodeHandle();
+            mPcPub = mNh.advertise<sensor_msgs::PointCloud2>("camera/left/points", 1);
+            mImuPub = mNh.advertise<sensor_msgs::Imu>("imu", 1);
+            image_transport::ImageTransport it{mNh};
+            mLeftImgPub = it.advertise("camera/left/image", 1);
+
             int resolution{};
             mPnh.param("grab_resolution", resolution, static_cast<std::underlying_type_t<sl::RESOLUTION>>(sl::RESOLUTION::HD720));
             mPnh.param("grab_target_fps", mGrabTargetFps, 50);
             mPnh.param("image_width", mImageWidth, 1280);
             mPnh.param("image_height", mImageHeight, 720);
-            mPnh.param("optimize_tag_detection", mDirectTagDetection, true);
+            mPnh.param("direct_tag_detection", mDirectTagDetection, false);
             std::string svoFile{};
             mPnh.param("svo_file", svoFile, {});
 
@@ -66,8 +55,8 @@ namespace mrover {
             sl::PositionalTrackingParameters positionalTrackingParameters;
             mZed.enablePositionalTracking(positionalTrackingParameters);
 
-            mGrabThread = std::thread(&ZedNode::grabUpdate, this);
-            mTagThread = std::thread(&ZedNode::tagUpdate, this);
+            mGrabThread = std::thread(&ZedNodelet::grabUpdate, this);
+            mTagThread = std::thread(&ZedNodelet::tagUpdate, this);
             //            if (optimizeTagDetection) {
             //                ROS_INFO("Loading tag detector nodelet");
             //                pluginlib::ClassLoader<TagDetectorNodelet> tagDetectorLoader{"mrover", "nodelet::Nodelet"};
@@ -79,68 +68,17 @@ namespace mrover {
         }
     }
 
-    void fillPointCloudMessage(sl::Mat& xyz, sl::Mat& bgr, sensor_msgs::PointCloud2Ptr const& msg, size_t tick) {
-        if (xyz.getWidth() != bgr.getWidth() || xyz.getHeight() != bgr.getHeight()) {
-            throw std::invalid_argument("XYZ and RGB images must be the same size");
-        }
-
-        auto imagePtr = bgr.getPtr<sl::uchar4>();
-        auto* pointCloudPtr = xyz.getPtr<sl::float4>();
-        //            auto* pointCloudNormalPtr = mPointCloudNormalMat.getPtr<sl::float4>();
-        msg->header.frame_id = "zed2i_left_camera_frame";
-        msg->header.seq = tick;
-        msg->header.stamp = ros::Time::now();
-        msg->is_bigendian = __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__;
-        msg->is_dense = false;
-        msg->height = xyz.getHeight();
-        msg->width = xyz.getWidth();
-        sensor_msgs::PointCloud2Modifier modifier{*msg};
-        modifier.setPointCloud2Fields(
-                8,
-                "x", 1, sensor_msgs::PointField::FLOAT32,
-                "y", 1, sensor_msgs::PointField::FLOAT32,
-                "z", 1, sensor_msgs::PointField::FLOAT32,
-                "rgb", 1, sensor_msgs::PointField::FLOAT32,
-                "normal_x", 1, sensor_msgs::PointField::FLOAT32,
-                "normal_y", 1, sensor_msgs::PointField::FLOAT32,
-                "normal_z", 1, sensor_msgs::PointField::FLOAT32,
-                "curvature", 1, sensor_msgs::PointField::FLOAT32);
-        auto* pointPtr = reinterpret_cast<Point*>(msg->data.data());
-        size_t size = msg->width * msg->height;
-        std::for_each(std::execution::par_unseq, pointPtr, pointPtr + size, [&](Point& point) {
-            size_t i = &point - pointPtr;
-            point.x = pointCloudPtr[i].x;
-            point.y = pointCloudPtr[i].y;
-            point.z = pointCloudPtr[i].z;
-            point.r = imagePtr[i].r;
-            point.g = imagePtr[i].g;
-            point.b = imagePtr[i].b;
-        });
-    }
-
-    void fillImageMessage(sl::Mat& bgr, sensor_msgs::Image& msg, size_t tick) {
-        msg.header.frame_id = "zed2i_left_camera_frame";
-        msg.header.seq = tick;
-        msg.header.stamp = ros::Time::now();
-        msg.height = bgr.getHeight();
-        msg.width = bgr.getWidth();
-        msg.encoding = sensor_msgs::image_encodings::BGRA8;
-        msg.step = bgr.getStepBytes();
-        msg.is_bigendian = __BYTE_ORDER__ == __ORDER_BIG_ENDIAN__;
-        auto* bgrPtr = bgr.getPtr<sl::uchar1>();
-        size_t size = msg.step * msg.height;
-        msg.data.resize(size);
-        std::memcpy(msg.data.data(), bgrPtr, size);
-    }
-
-    void ZedNode::tagUpdate() {
+    void ZedNodelet::tagUpdate() {
         try {
             if (!mDirectTagDetection) return;
 
             ROS_INFO("Starting tag thread");
-            mTagDetectorNode = boost::make_shared<TagDetectorNode>(mNh, mPnh, true);
 
-            std::this_thread::sleep_for(500ms);
+            // TODO: ugly, this prevents OpenCV fast alloc from crashing
+            std::this_thread::sleep_for(100ms);
+
+            mTagDetectorNode = boost::make_shared<TagDetectorNodelet>();
+            mTagDetectorNode->init("tag_detector", getRemappingArgs(), getMyArgv());
 
             // TODO: figure out why removing this causes a segfault
             cv::aruco::getPredefinedDictionary(cv::aruco::DICT_4X4_50);
@@ -160,7 +98,7 @@ namespace mrover {
         }
     }
 
-    void ZedNode::grabUpdate() {
+    void ZedNodelet::grabUpdate() {
         try {
             ROS_INFO("Starting grab thread");
             while (ros::ok()) {
@@ -237,29 +175,7 @@ namespace mrover {
                     mZed.getSensorsData(sensorData, sl::TIME_REFERENCE::CURRENT);
 
                     sensor_msgs::Imu imuMsg;
-                    imuMsg.header.stamp = ros::Time::now();
-                    imuMsg.header.seq = mUpdateTick;
-                    imuMsg.header.frame_id = "zed2i_imu_frame";
-                    imuMsg.orientation.x = sensorData.imu.pose.getOrientation().x;
-                    imuMsg.orientation.y = sensorData.imu.pose.getOrientation().y;
-                    imuMsg.orientation.z = sensorData.imu.pose.getOrientation().z;
-                    imuMsg.orientation.w = sensorData.imu.pose.getOrientation().w;
-                    for (int i = 0; i < 3; ++i)
-                        for (int j = 0; j < 3; ++j)
-                            imuMsg.orientation_covariance[i * 3 + j] = sensorData.imu.pose_covariance(i, j) * DEG2RAD * DEG2RAD;
-                    imuMsg.angular_velocity.x = sensorData.imu.angular_velocity.x * DEG2RAD;
-                    imuMsg.angular_velocity.y = sensorData.imu.angular_velocity.y * DEG2RAD;
-                    imuMsg.angular_velocity.z = sensorData.imu.angular_velocity.z * DEG2RAD;
-                    for (int i = 0; i < 3; ++i)
-                        for (int j = 0; j < 3; ++j)
-                            imuMsg.angular_velocity_covariance[i * 3 + j] = sensorData.imu.angular_velocity_covariance(i, j) * DEG2RAD * DEG2RAD;
-                    imuMsg.linear_acceleration.x = sensorData.imu.linear_acceleration.x;
-                    imuMsg.linear_acceleration.y = sensorData.imu.linear_acceleration.y;
-                    imuMsg.linear_acceleration.z = sensorData.imu.linear_acceleration.z;
-                    for (int i = 0; i < 3; ++i)
-                        for (int j = 0; j < 3; ++j)
-                            imuMsg.linear_acceleration_covariance[i * 3 + j] = sensorData.imu.linear_acceleration_covariance(i, j);
-
+                    fillImuMessage(sensorData.imu, imuMsg, mUpdateTick);
                     mImuPub.publish(imuMsg);
                 }
 
@@ -277,14 +193,10 @@ namespace mrover {
         }
     }
 
-    ZedNode::~ZedNode() {
+    ZedNodelet::~ZedNodelet() {
         ROS_INFO("ZED node shutting down");
         mTagThread.join();
         mGrabThread.join();
-    }
-
-    void ZedNodelet::onInit() {
-        dtl = boost::make_shared<ZedNode>(getMTNodeHandle(), getMTPrivateNodeHandle());
     }
 
 } // namespace mrover
@@ -292,7 +204,11 @@ namespace mrover {
 int main(int argc, char** argv) {
     ros::init(argc, argv, "zed_wrapper");
 
-    [[maybe_unused]] auto node = std::make_shared<mrover::ZedNode>();
+    // Start the ZED Nodelet
+    nodelet::Loader nodelet;
+    nodelet::M_string remap(ros::names::getRemappings());
+    nodelet::V_string nargv;
+    nodelet.load(ros::this_node::getName(), "mrover/ZedNodelet", remap, nargv);
 
     ros::spin();
