+++ src/perception/zed_wrapper/zed_wrapper.hpp
@@ -1,13 +1,15 @@
 #pragma once
 
 #include <condition_variable>
+#include <mutex>
+#include <thread>
 
 #include <sl/Camera.hpp>
 
-#include <image_transport/image_transport.h>
 #include <image_transport/publisher.h>
 #include <nodelet/nodelet.h>
 #include <ros/node_handle.h>
+#include <sensor_msgs/Imu.h>
 #include <sensor_msgs/PointCloud2.h>
 #include <tf2_ros/transform_broadcaster.h>
 #include <tf2_ros/transform_listener.h>
@@ -16,14 +18,13 @@
 
 namespace mrover {
 
-    class ZedNode {
+    class ZedNodelet : public nodelet::Nodelet {
     private:
         ros::NodeHandle mNh, mPnh;
 
         tf2_ros::Buffer mTfBuffer;
-        tf2_ros::TransformListener mTfListener;
+        tf2_ros::TransformListener mTfListener{mTfBuffer};
         tf2_ros::TransformBroadcaster mTfBroadcaster;
-        image_transport::ImageTransport mIt;
         ros::Publisher mPcPub, mImuPub;
         image_transport::Publisher mLeftImgPub;
 
@@ -47,30 +48,26 @@ namespace mrover {
         std::condition_variable mGrabDone;
         std::mutex mSwapPcMutex;
 
-        boost::shared_ptr<TagDetectorNode> mTagDetectorNode;
+        boost::shared_ptr<TagDetectorNodelet> mTagDetectorNode;
 
         size_t mUpdateTick = 0;
 
-    public:
-        ZedNode(ros::NodeHandle const& nh = {}, ros::NodeHandle const& pnh = {"~"});
+        void onInit() override;
 
-        ~ZedNode();
+    public:
+        ZedNodelet() = default;
+
+        ~ZedNodelet() override;
 
         void grabUpdate();
 
         void tagUpdate();
     };
 
-    class ZedNodelet : public nodelet::Nodelet {
-    public:
-        ZedNodelet() = default;
+    void fillPointCloudMessage(sl::Mat& xyz, sl::Mat& bgr, sensor_msgs::PointCloud2Ptr const& msg, size_t tick);
 
-        ~ZedNodelet() override = default;
+    void fillImageMessage(sl::Mat& bgr, sensor_msgs::Image& msg, size_t tick);
 
-    private:
-        void onInit() override;
-
-        boost::shared_ptr<ZedNode> dtl;
-    };
+    void fillImuMessage(sl::SensorsData::IMUData& imuData, sensor_msgs::Imu& msg, size_t tick);
 
 } // namespace mrover
